Move MultiValueMapping2Base into separate files.
